requirements.txt
Flask
Flask-Cors
requests
python-whois